refactor: rename dal to integrations
