+++ product/stream_processor/integrations/events/models/input.py
