+++ product/stream_processor/integrations/events/providers/eventbridge.py
@@ -5,12 +5,12 @@ import boto3
 import botocore.exceptions
 
 from product.constants import XRAY_TRACE_ID_ENV
-from product.stream_processor.dal.events.base import EventProvider
-from product.stream_processor.dal.events.constants import EVENTBRIDGE_PROVIDER_MAX_EVENTS_ENTRY
-from product.stream_processor.dal.events.exceptions import ProductNotificationDeliveryError
-from product.stream_processor.dal.events.functions import chunk_from_list
-from product.stream_processor.dal.events.models.input import Event
-from product.stream_processor.dal.events.models.output import EventReceipt, EventReceiptFail, EventReceiptSuccess
+from product.stream_processor.integrations.events.base import EventProvider
+from product.stream_processor.integrations.events.constants import EVENTBRIDGE_PROVIDER_MAX_EVENTS_ENTRY
+from product.stream_processor.integrations.events.exceptions import ProductNotificationDeliveryError
+from product.stream_processor.integrations.events.functions import chunk_from_list
+from product.stream_processor.integrations.events.models.input import Event
+from product.stream_processor.integrations.events.models.output import EventReceipt, EventReceiptFail, EventReceiptSuccess
 
 if TYPE_CHECKING:
     from mypy_boto3_events import EventBridgeClient
